+++ packages/kti_cqrs_rs/src/common/handler/event_handler.rs
@@ -1,0 +1,10 @@
+use std::error::Error;
+
+use async_trait::async_trait;
+
+#[async_trait]
+pub trait EventHandler: Send + Sync {
+  type Context;
+
+  async fn execute(&self, context: Self::Context) -> Result<(), Box<dyn Error>>;
+}