+++ packages/kti_cqrs_rs/src/common/handler/mod.rs
@@ -1,2 +1,3 @@
 pub mod command_handler;
+pub mod event_handler;
 pub mod query_handler;